allways extract scopes from repository
