+++ tests/Bootstrap.php
@@ -1,5 +1,11 @@
 <?php
 
 if (!@include_once __DIR__ . '/../vendor/autoload.php') {
-    exit("You must set up the project dependencies, run the following commands:\n> wget http://getcomposer.org/composer.phar\n> php composer.phar install\n");
+    $message = <<<MSG
+You must set up the project dependencies, run the following commands:
+> wget http://getcomposer.org/composer.phar
+> php composer.phar install
+MSG;
+
+    exit($message);
 }
