+++ tests/Stubs/ScopeEntity.php
@@ -1,20 +1,14 @@
 <?php
 
-namespace League\OAuth2\Server\Entities;
+namespace LeagueTests\Stubs;
 
 use League\OAuth2\Server\Entities\Interfaces\ScopeEntityInterface;
 use League\OAuth2\Server\Entities\Traits\EntityTrait;
 
-/**
- * Class ScopeEntity.
- */
 class ScopeEntity implements ScopeEntityInterface
 {
     use EntityTrait;
 
-    /**
-     * {@inheritdoc}
-     */
     public function jsonSerialize()
     {
         return $this->getIdentifier();
